+++ zcash_client_backend/src/fees/fixed.rs
@@ -10,6 +10,7 @@ use zcash_protocol::{
     value::{BalanceError, Zatoshis},
 };
 
+use crate::data_api::anchor_retention::PoolMigrationParams;
 use crate::data_api::{InputSource, wallet::TargetHeight};
 
 use super::{
@@ -104,6 +105,7 @@ impl<I: InputSource> ChangeStrategy for SingleOutputChangeStrategy<I> {
         params: &P,
         target_height: TargetHeight,
         anchor_height: BlockHeight,
+        zip318: &PoolMigrationParams,
         transparent_inputs: &[impl transparent::InputView],
         transparent_outputs: &[impl transparent::OutputView],
         sapling: &impl sapling_fees::BundleView<NoteRefT>,
@@ -141,6 +143,7 @@ impl<I: InputSource> ChangeStrategy for SingleOutputChangeStrategy<I> {
             #[cfg(feature = "orchard")]
             BundlePadding::DEFAULT,
             anchor_height,
+            zip318,
             self.change_memo.as_ref(),
             ephemeral_balance,
         )
@@ -149,6 +152,7 @@ impl<I: InputSource> ChangeStrategy for SingleOutputChangeStrategy<I> {
 
 #[cfg(test)]
 mod tests {
+    use crate::data_api::anchor_retention::{AnchorRetentionInterval, PoolMigrationParams};
     use ::transparent::bundle::TxOut;
     use zcash_primitives::transaction::fees::{
         fixed::FeeRule as FixedFeeRule, zip317::MINIMUM_FEE,
@@ -190,6 +194,7 @@ mod tests {
                 .unwrap()
                 .into(),
             BlockHeight::from_u32(1),
+            &PoolMigrationParams::new(AnchorRetentionInterval::ZIP_318),
             &[] as &[TestTransparentInput],
             &[] as &[TxOut],
             &(
@@ -234,6 +239,7 @@ mod tests {
                 .unwrap()
                 .into(),
             BlockHeight::from_u32(1),
+            &PoolMigrationParams::new(AnchorRetentionInterval::ZIP_318),
             &[] as &[TestTransparentInput],
             &[] as &[TxOut],
             &(